feat: add badge implementation - code smells + remove duplication
