+++ Showcase/Application/Utils/BaseImageCollectionViewController.swift
@@ -1,0 +1,26 @@
+//
+//  Vitamin iOS
+//  Apache License 2.0
+//
+
+import UIKit
+import Vitamin
+
+class BaseImageCollectionViewController: BaseHeaderedCollectionViewController {
+    var sections: [BaseImageSection] = []
+
+    override func collectionView(
+        _ collectionView: UICollectionView,
+        didSelectItemAt indexPath: IndexPath
+    ) {
+        UIPasteboard.general.string = sections[indexPath.section].items[indexPath.row].name
+    }
+}
+
+protocol BaseImageSection: BaseNamedSection {
+    var items: [BaseImageItem] { get }
+}
+
+protocol BaseImageItem {
+    var name: String { get }
+}
